cursor/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-dotenv==1.0
django-extensions==3.2.1
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
yarl==1.9.2
